Did assignment 25_break_encr

+++ 25_break_encr/breaker.c
@@ -16,7 +16,7 @@ size_t find_max(int * array){
 void decryptor(FILE *f){
   int array[26]={0};
   int c;
-  int j=0;
+  long int j=0;
   while((c=fgetc(f))!=EOF){
     if (isalpha(c)){
       c=c-'a';
